cleanup/common/db1_2py/msds_requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
Brotli==1.1.0
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
cryptography==46.0.2
cssselect2==0.8.0
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docling==2.61.2
docling-core==2.50.1
docling-ibm-models==3.10.2
docling-parse==4.7.1
docstring_parser==0.17.0
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.1
Faker==38.0.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gdown==5.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
instructor==1.13.0
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jpype1==1.6.0
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
langid==1.1.6
latex2mathml==3.78.1
lxml==6.0.2
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.4
ollama==0.6.0
omegaconf==2.3.0
openai==1.101.0
opencv-python==4.12.0.88
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
polyfactory==2.22.4
pre_commit==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.1
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg2==2.9.11
pure_eval==0.2.3
pycairo==1.28.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pylatexenc==2.10
pymongo==4.15.3
PyMuPDF==1.26.4
pypdf==6.1.3
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.1
rapidocr==3.4.2
referencing==0.37.0
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rlPyCairo==0.4.0
rpds-py==0.28.0
rtree==1.4.1
safetensors==0.6.2
scipy==1.16.3
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
svglib==1.6.0
sympy==1.14.0
tabula==1.0.5
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
ty==0.0.1a25
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
virtualenv==20.35.4
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
xhtml2pdf==0.2.17
xlsxwriter==3.2.9
yarl==1.22.0
zopfli==0.2.3.post1
